Move `ReactionType` unit tests into `reaction.rs`
